code clean up on errors
